+++ lib/middleman-akcms/series.rb
@@ -15,7 +15,12 @@ module Middleman::Akcms
         if res.data.series
           #name = res.data.series[:name] || res.category_resource.locals[:display_name]
           #number = res.data.series[:number] || nil
-          name = res.category_resource.try(:locals, :display_name) || File.dirname(res.path)
+          name =
+            begin
+              res.category_resource.locals[:display_name]
+            rescue
+              File.dirname(res.path)
+            end
           fname = File.split(res.path).last
           md = fname.match(/^([0-9]+)[_\-\s]/)
           number = (md.nil?) ? 0 : md[1].to_i
@@ -32,7 +37,6 @@ module Middleman::Akcms
         related_articles = modified_resources.select {|r| r.locals[:series][:name] == res.locals[:series][:name]}
         res.locals[:series][:related_articles] = related_articles
       }
-      #binding.pry
       used_resources + modified_resources
     end
     ################################################################